requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
flask
flask-cors